RecyclerView item prefetching
bug:27106058

If item inflation or binding are expensive, having them block the
production of a frame can lead to jank/stuttering as frames with such
work fail to meet rendering deadlines.

RecyclerView will now query LayoutManagers for views in advance of them
coming onscreen during scrolls and flings. These views are prefetched
into the Recycler cache during the UI thread's idle time between handing
a frame off to RenderThread, and the start of the next frame.

Change-Id: If3e485231243c188efd4cf3e512b9e8e13a50d47