MediaBrowserCompat: Make ItemCallback can take a null media item
When the implementation of MediaBrowserServiceCompat.onLoadItem returns null,
ItemCallback.onError was called, but the JavaDoc is saying that
ItemCallback.onItemLoaded can be called with a null result.

This CL makes ItemCallback.onItemLoaded be able to take a null result,
and ItemCallback.onError be called when the implementation doesn't override
the onLoadItem method or when the returned object is not MediaItem.

Bug: 27808084
Change-Id: Ifd25bcbd0775af221f3b48f715a8eb703caa77ad