Deprecate support-fragment's dependency on media-compat
Deprecate FragmentActivity's dependence on MediaControllerCompat and
therefore support-fragment's dependency on support-media-compat. This
will allow us to remove the compile time dependency in a future release.

New static methods in MediaControllerCompat are used to set the
Activity's MediaController. The addition of SupportActivity and its
ExtraData class allow MediaControllerCompat to store the
MediaControllerCompat object scoped to the Activity to
1) Support pre-API 21 devices
2) Prevent constant translation between framework and compat instances

Test: mmma frameworks/support, ./gradlew assemble, run SupportActivityTest

BUG: 30944164
Change-Id: I6db82487d7b6352bbcf2ab0202a7e1ffbd9af650