MediaRouter: Apply styles in chooser dialog based on isLightTheme
Unlike the controller dialog, colors of some components (background,
title) used in the chooser dialog is chosen based on the current theme,
while other components (icons, texts) are based on the isLightTheme
attribute. Therefore, when an application overrides the isLightTheme
attribute, the chooser dialog could become unusable.

This CL revisited the way of styling and theming of media router dialog
using theme overlay, so the appropriate theme overlay will be chosen and
applied based on the isLightTheme attribute.

In addition, this CL also includes the following:
  - Use a custom text view for title instead of setting framework
  dialog's title.
  - Make the chooser dialog inherit AppCompatDialog not Dialog.
  - When getting resources for views in the dialog, set the right
  Context.

Bug: 28040912, Bug: 30500649
Test: Tested with Support7Demo app and checked theme.
Change-Id: I4d5104e1231435f0b0721fd0af91255d0fff6afc