Restrict AVDC and VDC parsing to the current element instead of the whole document
This CL changes the AVDC xml parsing to parsing only within the current
element, which prevents AVDC parsing from always skipping to the end of
the doucment. So things that are defined after AVDC in the same document
can be picked up by the xml parser.

The same fix has been applied to VDC as well.

BUG: 31865175
Test: Manually following comment #1 in the bug above
Change-Id: I02cc2e6cd430ad53d9c01cd718231a943c4519e9