Preserve left+right compound drawables
Caused by the logic to remove the password dummy
compound drawable ran all the time, even when it
didn't need to. Due to the way that TextView returns
relative  drawables when absolutely set, this meant that
custom set compounds were wiped out.

Fixed making that logic only run when it needs to.
This still means that absolute set compounds are wiped
out when the password toggle is actually used. Added
documentation to explain.

Test: included
BUG: 32326780

Change-Id: I80e7c8f504c07bca437038167f6240dddf4bfe7f