Leanback: fix testSetSelectedPositionDetached
Leanback restores focus inside onLayoutChildren().
Due to the nature of leanback, it needs extra steps
to do the alignment inside layout. So there is no
need to run extra recycleView save/restore focus
logic.

This change "fixes" testSetSelectedPositionDetached on
TV, however the problem for phone/tablet still exists.
Because leanback always maintains a "selectedIndex".
When RV is detached and reattached, it will reselect
the same child. This is not true for a general RecyclerView
which does not maintain a "selectedIndex" and the
focused child will lose focus when detached.

This behavior can either be a WAI for general recyclerview. Or
recyclerView in general can behave similar to leanback
to maintain the selectedIndex.

Bug: 32673710
Test: GridWidgetTest
Change-Id: I05942c7c1c3e437d4a71c1f925e023280cd31b38