Fixed focus recovery when adapter changes during scrolling
When adapter changes during scrolling, consumePendingUpdateOperations
dispatched layout which didn't trigger ViewHolder updates only until
after the call to saveFocusInfo. Since in saveFocusInfo, we rely on
having an up-to-date state of viewholders after the pending update,
this created an issue when a focused item was removed when scrolling
was still in progress.
Also disabled the waitForIdleScroll in requestFocus to speed up the tests
in RecyclerViewFocusRecoveryTest.

Fixes: 32711883
Test: ./gradlew support-recyclerview-v7:connectedCheck
-Pandroid.testInstrumentationRunnerArguments.class=android.support.v7.widget.RecyclerViewFocusRecoveryTest

Change-Id: I26336cfa35cc9b42b6e38bed4f346495ddc10447