Fix night mode API for configuration-based color resources
API 21-24 have an issue where configuration based colors
are not purge from the drawable cache on configuration
changes. This has the side effect that colors which are
defined in values/ and values-night/ (as an example)
do not refresh when the night mode changes.

The fix included is beyond gross, but it's the only
method which works.

Test: included
BUG: 30132023

Change-Id: Id717cd443056be5cad11738a710219bfff62359a