Don't change visibility of fragment Views that aren't visible
Bug 32864127

When a fragment creates a view that is either invisible or gone,
it should remain that way. This means that the fragment won't have
any animations. When transactions are optimized, transitions
will avoid invisible and gone fragment views, but unoptimized
transactions will lead to failed transitions because none of the
views will be visible.

Test: gradlew connectedCheck in fragments/

Change-Id: Id5ab447a006cf37b30024bbe7efb3bb7d253f2ce