Make processors exception free
If a processor throws an exception, it halts all of the processing,
resulting in a lot of errors in javac output.
This CL changes room's processor behavior to just print errors and
continue as much as it can. This will eventually allow Room to
procude as much code as it can to avoid false positive erros in
javac output.

Bug: 32342709
Test: clean check passes w/o any changes
Change-Id: I54c9ad0cd64d88f8994d514479c9177fd6e598bb