Preparations for query adapter
This CL makes some changes in preparation for query adapters.

* re-packaged type adapters to be separated from query adapteres
* Changed TypeAdapter store to create chained TypeConverters that
can be merged with 1 ColumnAdapter to create a CompositeColumnAdapter.
In other words, now both CompositeColumnAdapter and
CompositeTypeConverter have just 2 items and they get chained as needed.
This simplifies the readability of our code a lot and has no effect on
the generated code.
* Implemented type converteres from pritivies to String. It is necessary
to create bind arguments for SQL.
* Cleaned up TypeAdapterStore

Test: BasicTypeConvertersTest.kt + existing tests
Bug: 32342709
Change-Id: Ibcd2e05796be2576847d6627b98d17339dfbea65