Switch mNestedRecyclerView to weakref
Fixes: 33278066
Test: new test, @SmallTests still pass

Previously, we cleared a nested RecyclerView to be safe when a holder
failed to recycle. However, this is a common occurrance with
animations, since they will hold onto the holders until they are done,
and return them at the end.

Instead, just keep the nested RecyclerView as a weak ref. Probably
unnecessary, but maintains the guarantee that if someone removes or
swaps an inner RecyclerView (note: a gross case), we won't leak it.

Change-Id: I3ee97608f8fb3687f17d861438bc4bb895d27a65