@Delete support
This CL adds support for abstract methods in DAO annoated with
@Delete.

It is fairly similar to Insert but there are subtle changes so
they share only limited number of classes. I've decided to name
these methods ShortcutMethods (since they are shortcuts to writing
SQL). Suggestions are welcome.

There will also be support for @Update but the CL is getting big
so i decided to spare it for another CL.

Bug: 32342709
Test: DaoWriterTest, SimpleEntityReadWriteTest
Test: DeletionMethodProcessorTest
Change-Id: Ifbf69067b014f1ae5349567a74b4d3f9efb846c7