Backport IME_FLAG_NO_PERSONALIZED_LEARNING flag
This CL backports a new bit flag for EditorInfo#imeOptions.

Since IME_FLAG_NO_PERSONALIZED_LEARNING is designed to be an optional
flag from applications to IMEs that is never used by the Android OS,
just defining the same constant in EditorInfoCompat is good enough
to use this flag on existing devices for faster adaptation of this
flag.

Bug: 28157942
Test: Compile
Change-Id: Idc5bb2b85c3776f3736b7c9717f5ed40f2b6304e