Backport EditorInfo.IME_FLAG_FORCE_ASCII
This CL backports EditorInfo#IME_FLAG_FORCE_ASCII which was added in
API 16.

Since IME_FLAG_FORCE_ASCII is designed to be an optional flag from
applications to IMEs that is never used by the Android OS, just
defining the same constant in EditorInfoCompat is good enough to use
this flag on existing devices for faster adaptation of this flag.

Bug: 34141636
Test: Compile
Change-Id: I573095cad260c501b77b427e063736779a7c952f