Database invalidation tracking
This CL introduces updated table tracking via sqlite triggers.

There is a new InvalidationTracker class, which creates a version
table in memory that keeps a version for each table. It also
creates triggers for each table and these triggers update the
version table when observed table changes.

When a transaction is committed, we trigger the InvalidationTracker
to refresh itself. This covers all queries that go through Room
but if someone grabs the db and runs a query directly, it wont
work. To support them, there is a public refresh method in the
InvalidationTracker. We may also work around it by wrapping the
SupportSQLiteDatabase.

In a followup CL, I'll introduce an implementataion of
LiveData<T> that observes this table automatically.

Bug: 32342709
Test: InvalidationTrackerTest
Change-Id: I9d0a0cc9dc537200f3fd3cd229df9bfeaf52b56d