Introduce forked context
This CL introduces forkable context which allows having scoped
changes in the processer.
A Context now can be forked per element and all ignored annotations
are parsed during the fork which automatically scopes them.
Each processor has to fork the context, and if they don't, they
won't have scoped changes but they won't be able to mess up
with parent's context since it is immutable.

I've changed each processor to be 1 time use so that they can be
scoped to their element. This means they receive all parameters
in the constructor (rather than receiving the element in the
process method). This churns more memory but who cares, it is JVM.

This new scope allows us to get rid of report errors methods in
column adapters because they can freely report their warnings
and filtering happens on the logger side.

I'm hoping that implementing TypeConverters on top of this will be
more straightforward but saving it for another CL.

Bug: 32342709
Test: all existing tests pass after api modifications
Change-Id: Ie6b6a6dd83149c561465905faf5438e006440460