Fix package*ClassFilesForCoverageReport tasks.
jarifyTask had an implicit dependency on actual java compilation
that happened to just work before we enabled parallel builds. Once
we did that, sometimes jarifyTask would run before compilation was
over. This change fixes that.

Additionally, this CL starts excluding generated BuildConfig that
should not be included in coverage reports.

Bug: 34932818
Test: rm -fr ../../out/ && ./gradlew support-compat:packageDebugAndroidTestClassFilesForCoverageReport
Change-Id: I8fbbd75d98b0ee95ae9f29013f4a699c71f35068